fix(app): Reload required queries after snapshot creation
